Move util files from /packs to /src
