+++ app/javascript/src/Feed.jsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from 'react';
 import ReactDOM from 'react-dom';
-import { getTweets, postTweet, deleteTweet } from '../packs/requests';
+import { getTweets, postTweet, deleteTweet } from './requests';
 import Navbar from './navbar';
-import { getCurrentUser, countUsersTweets } from '../packs/utils';
+import { getCurrentUser, countUsersTweets } from './utils';
 import Footer from './footer';
 
 const Feed = () => {
