+++ app/javascript/src/home.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import ReactDOM from 'react-dom';
 import $ from 'jquery';
-import { createUser, logInUser, authenticateUser } from '../packs/requests';
+import { createUser, logInUser, authenticateUser } from './requests';
 import Navbar from './navbar';
 import Footer from './footer';
 import './home.scss';
